% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptreject_DAG.R
\name{acceptreject_DAG}
\alias{acceptreject_DAG}
\title{Accept/Reject a proposed DAG}
\usage{
acceptreject_DAG(
  X,
  tXX,
  currentDAG,
  proposedDAG,
  node,
  op.type,
  a,
  U,
  w,
  current.opcard,
  proposed.opcard
)
}
\arguments{
\item{X}{Matrix of centered data}

\item{tXX}{Empirical covariance matrix (given centered data)}

\item{currentDAG}{As the name suggests}

\item{proposedDAG}{As the name suggests}

\item{node}{Nodes that differ between the two DAGs}

\item{op.type}{The operation performed on the current DAG to obtain the proposed DAG}

\item{a}{hyperparameter}

\item{U}{hyperparameter}

\item{w}{hyperparameter}
}
\value{
A Boolean indicating whether the proposed DAG has been accepted or not
}
\description{
Accept/Reject a proposed DAG
}
