% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_graphNEL.R
\name{as_graphNEL}
\alias{as_graphNEL}
\title{Transform matrix into graphNEL object}
\usage{
as_graphNEL(DAG)
}
\arguments{
\item{DAG}{Adjacency matrix of a DAG}
}
\value{
A graphNEL object
}
\description{
Internal function to transform an adjacency matrix into a graphNEL object, useful for plotting
}
\examples{
# Randomly generate DAG
q <- 4; w = 0.2
set.seed(123)
DAG <- rDAG(q,w)
as_graphNEL(DAG)
}
