% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd_encode.R
\name{bd_encode}
\alias{bd_encode}
\title{Convert matrix into strings (internal function)}
\usage{
bd_encode(matrix, separator = ";")
}
\arguments{
\item{matrix}{a matrix to convert into string}

\item{separator}{symbol used to separate elements of the matrix in the string}
}
\description{
This function converts matrices into string objects.
It is implemented in \code{learn_DAG} when \code{save.memory = TRUE}
}
