% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causaleffect.R
\name{causaleffect}
\alias{causaleffect}
\title{Estimate the causal effect given a DAG and associated Cholesky parameters}
\usage{
causaleffect(targets, response, DAG, L, D)
}
\arguments{
\item{targets}{vector, targets of intervention}

\item{response}{numeric, response variable}

\item{DAG}{matrix, DAG needed to estimate the causal effect}

\item{L}{matrix, first Cholesky parameter}

\item{D}{matrix, second Cholesky parameter}
}
\value{
a vector of length |targets|
}
\description{
Estimate the causal effect given a DAG and associated Cholesky parameters
}
