% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DW_nodemll.R
\name{DW_nodemll}
\alias{DW_nodemll}
\title{Ok controllare det matrice una dimensione (dà errore)}
\usage{
DW_nodemll(node, DAG, X, tXX, a, U)
}
\arguments{
\item{U}{}
}
\value{

}
\description{
Ok controllare det matrice una dimensione (dà errore)
}
