% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa}
\alias{fa}
\title{Learn the family of a node given a DAG}
\usage{
fa(node, DAG)
}
\arguments{
\item{node}{Questa e' una prova di commento}

\item{DAG}{}
}
\value{
fa
}
\description{
Learn the family of a node given a DAG
}
