% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opcard.R
\name{get_opcard}
\alias{get_opcard}
\title{Build the set of direct successors of a DAG (internal function)}
\usage{
get_opcard(DAG)
}
\arguments{
\item{DAG}{The DAG for which we want to list the set of direct successors}
}
\description{
Build the set of direct successors of a DAG (internal function)
}
