% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_causaleffect.R
\name{learn_causaleffect}
\alias{learn_causaleffect}
\title{Estimate (joint) causal effects from DAG learning procedure}
\usage{
learn_causaleffect(learnDAG_output, targets, response, BMA = FALSE)
}
\arguments{
\item{learnDAG_output}{list, output of learn_DAG()}

\item{targets}{vector, targets of intervention}

\item{response}{numeric, response variable}

\item{BMA}{if TRUE, a single vector containing a Bayesian model averaging
estimate of the causal effect is returned}
}
\value{
Either a vector containing BMA causal effect estimate, or a matrix with G rows containing different
causal effects estimates (one for each DAG in the mcmc)
}
\description{
Estimate (joint) causal effects from DAG learning procedure
}
