% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_dag.R
\name{learn_DAG}
\alias{learn_DAG}
\title{Learn an unknown DAG from data using a PAS algorithm}
\usage{
learn_DAG(G, burnin, data, a, U, w, fast = FALSE, save.memory = FALSE)
}
\arguments{
\item{G}{Desired sample size}

\item{burnin}{Burnin}

\item{data}{the data}

\item{a}{hyperparameter}

\item{U}{hyperparameter}

\item{w}{hyperparameter}

\item{thin}{thinning}
}
\value{

}
\description{
Learn an unknown DAG from data using a PAS algorithm
}
