% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_edgeprobs.R
\name{learn_edgeprobs}
\alias{learn_edgeprobs}
\title{Estimate posterior probabilities of edge inclusion}
\usage{
learn_edgeprobs(learnDAG_output)
}
\arguments{
\item{learnDAG_output}{output of learn_dag()}
}
\value{
Matrix containing probability of edge inclusion
}
\description{
Estimate posterior probabilities of edge inclusion
}
