% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MAPdag.R
\name{learn_MAPdag}
\alias{learn_MAPdag}
\title{Maximum a posteriori DAG estimation from mcmc output}
\usage{
learn_MAPdag(learnDAG_output)
}
\arguments{
\item{learnDAG_output}{output of learn_dag()}
}
\value{
Adjacency matrix of MAP DAG
}
\description{
Maximum a posteriori DAG estimation from mcmc output
}
