% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MPMdag.R
\name{learn_MPMdag}
\alias{learn_MPMdag}
\title{Median Probability DAG estimation}
\usage{
learn_MPMdag(learnDAG_output)
}
\arguments{
\item{learnDAG_output}{output of learn_dag()}
}
\value{
Adjacency matrix of Median probability DAG
}
\description{
Median Probability DAG estimation
}
