% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bcdag.R
\name{new_bcdag}
\alias{new_bcdag}
\title{Create new bcdag objects}
\usage{
new_bcdag(x = list(), input = list(), type = "complete")
}
\arguments{
\item{x}{A list}

\item{input}{A list containing the inputs given to \code{learn_DAG()}}

\item{type}{A string indicating whether the output produced by \code{learn_DAG()} should be of type "complete", "compressed", "collapsed" or "compressed and collapsed"}
}
\value{
An S3 object of class \code{bcdag}
}
\description{
Internal function used as constructor for S3 objects of class \code{bcdag}, which constitute the output of function \code{learn_DAG()} and the input of functions belonging to the get_ family such as \code{get_causaleffect()}.
}
\references{
F. Castelletti and A. Mascaro (2022). BCDAG: An R package for Bayesian structural and Causal learning of Gaussian DAGs. \emph{arXiv pre-print}, url: https://arxiv.org/abs/2201.12003
}
\author{
Federico Castelletti and Alessandro Mascaro
}
