% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_bcdagCE.R
\name{new_bcdagCE}
\alias{new_bcdagCE}
\title{Create new bcdagCE objects}
\usage{
new_bcdagCE(x = list(), input = list(), type = "complete")
}
\arguments{
\item{x}{A list}

\item{input}{A list containing the inputs given to \code{learn_DAG()} and \code{get_causaleffect()}}

\item{type}{A string indicating whether the output produced by \code{learn_DAG()} was of type "complete", "compressed", "collapsed" or "compressed and collapsed"}
}
\value{
An S3 object of class \code{bcdagCE}
}
\description{
Internal function used as constructor for S3 objects of class \code{bcdagCE},
which constitute the output of function \code{get_causaleffect()}.
}
\references{
F. Castelletti and A. Mascaro (2022). BCDAG: An R package for Bayesian structural and Causal learning of Gaussian DAGs. \emph{arXiv pre-print}, url: https://arxiv.org/abs/2201.12003
}
\author{
Federico Castelletti and Alessandro Mascaro
}
