% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{operation}
\alias{operation}
\title{A simple function to perform an add/remove/reverse edge operation on a DAG}
\usage{
operation(op, A, nodes)
}
\arguments{
\item{op}{The type of operation performed: op == 1 corresponds to adding an edge, op == 2 corresponds to removing an edge, op == 3 corresponds to reversing an edge}

\item{A}{The adjacency matrix of the DAG on which the operation is performed}

\item{nodes}{The nodes on which the operation is to be performed}
}
\value{
The adjacency matrix of the modified DAG
}
\description{
A simple function to perform an add/remove/reverse edge operation on a DAG
}
