% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa.R
\name{pa}
\alias{pa}
\title{Find the parents of a node in a DAG (internal function)}
\usage{
pa(node, DAG)
}
\arguments{
\item{node}{numerical label of the node in \code{DAG}}

\item{DAG}{\eqn{(q,q)} adjacency matrix of the DAG}
}
\value{
A numerical vector with the labels of the parents of \code{node} in \code{DAG}
}
\description{
This function finds the set of parents of \code{node} in \code{DAG}
}
