% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propose_DAG.R
\name{propose_DAG}
\alias{propose_DAG}
\title{An internal function that randomly modifies a DAG and checks that the resulting DAG is a DAG}
\usage{
propose_DAG(DAG, fast)
}
\arguments{
\item{DAG}{Adjacency matrix of the current DAG}
}
\value{
A list containing the proposed DAG, the type of operation performed, the nodes on which the operation has been performed and the cardinality of all the possible operations
}
\description{
An internal function that randomly modifies a DAG and checks that the resulting DAG is a DAG
}
