% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDAGWishart.R
\name{rDAGWishart}
\alias{rDAGWishart}
\title{Random samples from a DAG-Wishart distribution with G&H hyperparameters}
\usage{
rDAGWishart(n, DAG, a, U)
}
\arguments{
\item{U}{}
}
\value{

}
\description{
Random samples from a DAG-Wishart distribution with G&H hyperparameters
}
