% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnodeDAGWishart.R
\name{rnodeDAGWishart}
\alias{rnodeDAGWishart}
\title{Sample 1 observation for a single node of a DAG from a DAG-Wishart distribution}
\usage{
rnodeDAGWishart(node, DAG, aj, U)
}
\arguments{
\item{node}{The node for which}

\item{U}{}
}
\value{

}
\description{
Sample 1 observation for a single node of a DAG from a DAG-Wishart distribution
}
