% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{fa}
\alias{fa}
\title{Find the family of a node in a DAG (internal function)}
\usage{
fa(node, DAG)
}
\arguments{
\item{node}{numerical label of the node in \code{DAG}}

\item{DAG}{\eqn{(q,q)} adjacency matrix of the DAG}
}
\value{
A numerical vector with the labels of the family of \code{node} in \code{DAG}
}
\description{
This function finds the family (union of \code{node} and its parents) of \code{node} in \code{DAG}
}
