% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opcard.R
\name{get_opcard}
\alias{get_opcard}
\title{Find the direct successors DAGs of an input DAG (internal function)}
\usage{
get_opcard(DAG)
}
\arguments{
\item{DAG}{The input DAG}
}
\description{
Find the direct successors DAGs of an input DAG (internal function)
}
