% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation.R
\name{operation}
\alias{operation}
\title{Perform local moves given a DAG (internal function)}
\usage{
operation(op, A, nodes)
}
\arguments{
\item{op}{numerical type in \eqn{{1,2,3}} of the operator applied to \code{DAG}}

\item{A}{\eqn{(q,q)} adjacency matrix of the input DAG}

\item{nodes}{numerical labels of nodes on which the operator is applied, a \eqn{(2,1)} vector}
}
\value{
The \eqn{(q,q)} adjacency matrix of the modified DAG
}
\description{
This function locally modifies a DAG by inserting (\code{op = 1}), deleting (\code{op = 2}) or reversing (\code{op = 3}) an edge between two \code{nodes}
}
